+++ philo/srcs/time.c
@@ -6,7 +6,7 @@
 /*   By: dvan-kri <dvan-kri@student.codam.nl>         +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2022/04/21 15:10:27 by dvan-kri      #+#    #+#                 */
-/*   Updated: 2022/05/23 13:43:08 by dvan-kri      ########   odam.nl         */
+/*   Updated: 2022/05/23 15:21:26 by dvan-kri      ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,10 +15,10 @@
 #include "../incl/philosophers.h"
 #include "../incl/time.h"
 
-void	usleep_more_accurate(t_data *data, size_t time)
+void	usleep_more_accurate(t_data *data, long long time)
 {
-	size_t			time_passed;
-	size_t			time_start;
+	long long		time_passed;
+	long long		time_start;
 	int				times;
 
 	times = 0;
@@ -47,17 +47,17 @@ void	set_start_time(t_data *data)
 	data->start_time = time_stamp();
 }
 
-size_t	time_stamp(void)
+long long	time_stamp(void)
 {
 	struct timeval	the_time;
-	size_t			time_in_milliseconds;
+	long long			time_in_milliseconds;
 
 	gettimeofday(&the_time, NULL);
 	time_in_milliseconds = (the_time.tv_sec * 1000) + (the_time.tv_usec / 1000);
 	return (time_in_milliseconds);
 }
 
-size_t	time_passed(size_t start)
+long long	time_passed(long long start)
 {
 	return (time_stamp() - start);
 }
